cis-backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
groq>=1.1.0
wikipedia-api>=0.7.1
networkx>=3.3
python-dotenv>=1.0.0
datasets>=3.0.0
scipy>=1.14.0
pydantic>=2.9.0
transformers>=4.46.0
torch>=2.3.0
